% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-args.R
\name{check_arg}
\alias{check_arg}
\title{Verify that the argument `arg` is of the correct class, length, and
Verify that the argument `arg` is of the correct class, length, and
within in a set of values}
\usage{
check_arg(
  arg = NULL,
  chk_class = NULL,
  chk_dim = NULL,
  chk_len = NULL,
  chk_is_in = NULL,
  allow_null = FALSE
)
}
\arguments{
\item{arg}{The object to check}

\item{chk_class}{A character string of the name of the class to ensure `arg` is}

\item{chk_dim}{A vector of the dimensions to ensure `arg` is the same dimensions as}

\item{chk_len}{A numeric value to ensure `arg` has the length of}

\item{chk_is_in}{A vector of values to ensure `arg` is in}

\item{allow_null}{If `TRUE` the argument is allowed to be `NULL`}
}
\value{
TRUE, invisibly or the function will throw an error if `arg` does not follow
the constraints given; FALSE is not returned
}
\description{
Verify that the argument `arg` is of the correct class, length, and
Verify that the argument `arg` is of the correct class, length, and
within in a set of values
}
\examples{
\dontrun{
check_arg(23, "numeric", 1) # Succeeds
check_arg(23, "numeric", 1, 1:20) # Fails
}
}
