% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-args.R
\name{get_args}
\alias{get_args}
\title{Get a named list of the argument values in any function call}
\usage{
get_args()
}
\value{
A named list of the argument values used in a function call
}
\description{
Get a named list of the argument values in any function call
}
